fix code

+++ services/context/inc/widget_context.h
@@ -103,7 +103,7 @@ private:
         AuthType authType { 0 };
         uint64_t credentialDigest;
         uint16_t credentialCount;
-        int32_t pinExpiredInfo;
+        int64_t pinExpiredInfo;
     };
 
     uint64_t contextId_ {0};
