+++ services/core/inc/authentication.h
@@ -48,7 +48,7 @@ public:
         uint16_t credentialCount{0};
         int32_t sdkVersion{0};
         int32_t userId;
-        int32_t pinExpiredInfo;
+        int64_t pinExpiredInfo;
     };
     struct AuthExecutorMsg {
         uint64_t executorIndex;
